requirements.txt
absl-py==2.0.0
attrs==23.1.0
cffi==1.16.0
contourpy==1.2.0
cx_Freeze==6.15.16
cx_Logging==3.2.0
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.46.0
kiwisolver==1.4.5
lief==0.14.1
matplotlib==3.8.2

MouseInfo==0.1.3
numpy==1.26.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
protobuf==3.20.3
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
pytweening==1.0.7
six==1.16.0
sounddevice==0.4.6
